requirements.txt
netlas==0.6.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
